backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "bedrock-agentcore>=1.4.4",
    "strands-agents>=1.29.0",
]
